+++ test/functional/p2p_segwit.py
@@ -217,6 +217,9 @@ class SegWitTest(BitcoinTestFramework):
         add_witness_commitment(block, nonce)
         block.solve()
 
+    def get_segwit_info(self, node_id):
+        return self.nodes[node_id].getforkinfo()['segwit']
+
     def run_test(self):
         # Setup the p2p connections
         # self.test_node sets NODE_WITNESS|NODE_NETWORK
@@ -285,12 +288,14 @@ class SegWitTest(BitcoinTestFramework):
             self.log.info("Subtest: {} (Segwit status = {})".format(func.__name__, self.segwit_status))
             # Assert segwit status is as expected
             assert_equal(get_bip9_status(self.nodes[0], 'segwit')['status'], self.segwit_status)
+            assert_equal(self.get_segwit_info(0)['active'], self.segwit_status == 'active')
             func(self, *args, **kwargs)
             # Each subtest should leave some utxos for the next subtest
             assert self.utxo
             self.sync_blocks()
             # Assert segwit status is as expected at end of subtest
             assert_equal(get_bip9_status(self.nodes[0], 'segwit')['status'], self.segwit_status)
+            assert_equal(self.get_segwit_info(0)['active'], self.segwit_status == 'active')
 
         return func_wrapper
 
@@ -710,8 +715,10 @@ class SegWitTest(BitcoinTestFramework):
         height = self.nodes[0].getblockcount()
         self.nodes[0].generate(VB_PERIOD - (height % VB_PERIOD) - 2)
         assert_equal(get_bip9_status(self.nodes[0], 'segwit')['status'], 'locked_in')
+        assert_equal(self.get_segwit_info(0)['active'], False)
         self.nodes[0].generate(1)
         assert_equal(get_bip9_status(self.nodes[0], 'segwit')['status'], 'active')
+        assert_equal(self.get_segwit_info(0)['active'], True)
         self.segwit_status = 'active'
 
     @subtest