+++ test/functional/test_framework/util.py
@@ -343,8 +343,8 @@ def delete_cookie_file(datadir):
         os.remove(os.path.join(datadir, "regtest", ".cookie"))
 
 def get_bip9_status(node, key):
-    info = node.getblockchaininfo()
-    return info['bip9_softforks'][key]
+    info = node.getforkinfo()
+    return info[key]['bip9']
 
 def set_node_times(nodes, t):
     for node in nodes: